♻️ A20GQR context: decompose wiki command
