+++ .agentplane/tasks/202605282336-A20GQR/pr/github-body.md
@@ -1,0 +1,33 @@
+Task: `202605282336-A20GQR`
+Title: Context wiki command decomposition
+Canonical task record: `.agentplane/tasks/202605282336-A20GQR/README.md`
+
+## Summary
+
+Context wiki command decomposition
+
+Decompose packages/agentplane/src/commands/context/wiki.ts by extracting wiki page rendering and lint/catalog helpers while preserving context wiki command behavior.
+
+## Scope
+
+- In scope: Decompose packages/agentplane/src/commands/context/wiki.ts by extracting wiki page rendering and lint/catalog helpers while preserving context wiki command behavior.
+- Out of scope: unrelated refactors not required for "Context wiki command decomposition".
+
+## Verification
+
+- State: pending
+- Note: Not recorded yet.
+- Canonical workflow state lives in the task README.
+
+<details>
+<summary>Raw evidence</summary>
+
+- Updated: 2026-05-28T23:36:39.147Z
+- Branch: task/202605282336-A20GQR/context-wiki-decomposition
+- Head: computed live by `agentplane pr check` / `agentplane integrate`
+
+```text
+No changes detected.
+```
+
+</details>
