+++ .agentplane/tasks/202605282336-A20GQR/pr/review.md
@@ -1,0 +1,36 @@
+# PR Review
+
+Created: 2026-05-28T23:36:39.147Z
+
+## Task
+
+- Task: `202605282336-A20GQR`
+- Title: Context wiki command decomposition
+- Status: DOING
+- Branch: `task/202605282336-A20GQR/context-wiki-decomposition`
+- Canonical task record: `.agentplane/tasks/202605282336-A20GQR/README.md`
+
+## Verification
+
+- State: pending
+- Note: Not recorded yet.
+- Canonical workflow state lives in the task README.
+
+## Handoff Notes
+
+- No handoff notes recorded yet. Use `agentplane pr note ...` to append one.
+
+<!-- BEGIN AUTO SUMMARY -->
+<details>
+<summary>Raw evidence</summary>
+
+- Updated: 2026-05-28T23:36:39.147Z
+- Branch: task/202605282336-A20GQR/context-wiki-decomposition
+- Head: computed live by `agentplane pr check` / `agentplane integrate`
+
+```text
+No changes detected.
+```
+
+</details>
+<!-- END AUTO SUMMARY -->
